+++ backend/packages/integrations/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @cat-factory/integrations
 
+## 0.123.6
+
+### Patch Changes
+
+- Updated dependencies [2580fee]
+- Updated dependencies [eb4ca17]
+  - @cat-factory/kernel@0.233.0
+  - @cat-factory/contracts@0.231.0
+
 ## 0.123.5
 
 ### Patch Changes